chore(algokit_transact): more adjustments to match asset config patterns
