+++ crates/algokit_transact/src/tests.rs
@@ -6,14 +6,13 @@ use crate::{
         AddressMother, TransactionGroupMother, TransactionHeaderMother, TransactionMother,
     },
     transactions::FeeParams,
-    Address, AlgorandMsgpack, ApplicationCallTransactionBuilder, BoxReference,
-    EstimateTransactionSize, OnApplicationComplete, SignedTransaction, Transaction, TransactionId,
-    Transactions,
+    Address, AlgorandMsgpack, EstimateTransactionSize, SignedTransaction, Transaction,
+    TransactionId, Transactions,
 };
 use base64::{prelude::BASE64_STANDARD, Engine};
 use pretty_assertions::assert_eq;
 
-fn check_transaction_encoding(tx: &Transaction, expected_encoded_len: usize) {
+pub fn check_transaction_encoding(tx: &Transaction, expected_encoded_len: usize) {
     let encoded = tx.encode().unwrap();
     let decoded = Transaction::decode(&encoded).unwrap();
     assert_eq!(decoded, *tx);
@@ -36,7 +35,7 @@ fn check_transaction_encoding(tx: &Transaction, expected_encoded_len: usize) {
     assert_eq!(encoded.len(), expected_encoded_len);
 }
 
-fn check_signed_transaction_encoding(
+pub fn check_signed_transaction_encoding(
     tx: &Transaction,
     expected_encoded_len: usize,
     auth_address: Option<Address>,
@@ -52,7 +51,7 @@ fn check_signed_transaction_encoding(
     assert_eq!(decoded_stx, signed_tx);
 }
 
-fn check_transaction_id(tx: &Transaction, expected_tx_id: &str) {
+pub fn check_transaction_id(tx: &Transaction, expected_tx_id: &str) {
     let signed_tx = SignedTransaction {
         transaction: tx.clone(),
         signature: Some([0; ALGORAND_SIGNATURE_BYTE_LENGTH]),
@@ -410,235 +409,4 @@ fn test_signed_transaction_group_encoding() {
         assert_eq!(encoded_signed_tx, signed_grouped_tx.encode().unwrap());
         assert_eq!(decoded_signed_tx, signed_grouped_tx);
     }
-}
-
-#[test]
-fn test_application_create_transaction_encoding() {
-    let application_create_tx = TransactionMother::application_create().build().unwrap();
-
-    check_transaction_id(
-        &application_create_tx,
-        "L6B56N2BAXE43PUI7IDBXCJN5DEB6NLCH4AAN3ON64CXPSCTJNTA",
-    );
-    check_transaction_encoding(&application_create_tx, 1386);
-}
-
-#[test]
-fn test_application_call_encoding() {
-    let application_call_tx = TransactionMother::application_call().build().unwrap();
-
-    check_transaction_id(
-        &application_call_tx,
-        "6Y644M5SGTKNBH7ZX6D7QAAHDF6YL6FDJPRAGSUHNZLR4IKGVSPQ",
-    );
-    check_transaction_encoding(&application_call_tx, 377);
-}
-
-#[test]
-fn test_application_update_encoding() {
-    let application_update_tx = TransactionMother::application_update().build().unwrap();
-
-    check_transaction_id(
-        &application_update_tx,
-        "NQVNJ5VWEDX42DMJQIQET4QPNUOW27EYIPKZ4SDWKOOEFJQB7PZA",
-    );
-    check_transaction_encoding(&application_update_tx, 7069);
-}
-
-#[test]
-fn test_application_delete_transaction_encoding() {
-    let application_delete_tx = TransactionMother::application_delete().build().unwrap();
-
-    check_transaction_id(
-        &application_delete_tx,
-        "XVVC7UDLCPI622KCJZLWK3SEAWWVUEPEXUM5CO3DFLWOBH7NOPDQ",
-    );
-    check_transaction_encoding(&application_delete_tx, 263);
-}
-
-#[test]
-fn test_application_opt_in_transaction_encoding() {
-    let application_opt_in_tx = TransactionMother::application_opt_in().build().unwrap();
-
-    check_transaction_id(
-        &application_opt_in_tx,
-        "BNASGY47TXXUTFUZPDAGGPQKK54B4QPEEPDTJIZFDXC64WQH4GOQ",
-    );
-    check_transaction_encoding(&application_opt_in_tx, 247);
-}
-
-#[test]
-fn test_application_close_out_transaction_encoding() {
-    let application_close_out_tx = TransactionMother::application_close_out().build().unwrap();
-
-    check_transaction_id(
-        &application_close_out_tx,
-        "R4LXOUN4KPRIILRLIYKMA2DJ4HKCXWCD5TYWGH76635KUHGFNTUQ",
-    );
-    check_transaction_encoding(&application_close_out_tx, 131);
-}
-
-#[test]
-fn test_application_clear_state_transaction_encoding() {
-    let application_clear_state_tx = TransactionMother::application_clear_state()
-        .build()
-        .unwrap();
-
-    check_transaction_id(
-        &application_clear_state_tx,
-        "XQE2YKONC62QXSXDIRJ7CL6YDWP45JXCQO6N7DAAFQH7DJM6BEKA",
-    );
-    check_transaction_encoding(&application_clear_state_tx, 131);
-}
-
-#[test]
-fn test_0_box_ref_application_call_transaction_encoding() {
-    let application_call_tx = TransactionMother::application_call_example()
-        .box_references(vec![BoxReference {
-            app_id: 0,
-            name: "b1".as_bytes().to_vec(),
-        }])
-        .build()
-        .unwrap();
-
-    check_transaction_id(
-        &application_call_tx,
-        "LXUGSM4264PQ2YSSO3JW535NHGC5JESKLQS6ITONGO2S6ATEWM2A",
-    );
-    check_transaction_encoding(&application_call_tx, 138);
-}
-
-#[test]
-fn test_app_id_box_ref_application_call_transaction_encoding() {
-    let application_call_tx = TransactionMother::application_call_example()
-        .box_references(vec![BoxReference {
-            app_id: 12345,
-            name: "b1".as_bytes().to_vec(),
-        }])
-        .build()
-        .unwrap();
-
-    check_transaction_id(
-        &application_call_tx,
-        "LXUGSM4264PQ2YSSO3JW535NHGC5JESKLQS6ITONGO2S6ATEWM2A",
-    );
-
-    let encoded = application_call_tx.encode().unwrap();
-    let decoded = Transaction::decode(&encoded).unwrap();
-
-    if let Transaction::ApplicationCall(decoded_app_call) = decoded {
-        assert_eq!(
-            decoded_app_call.box_references.as_ref().unwrap()[0].app_id,
-            0
-        );
-    } else {
-        panic!("Expected ApplicationCall transaction type");
-    }
-}
-
-#[test]
-fn test_external_box_refs_application_call_transaction_encoding() {
-    let application_call_tx = TransactionMother::application_call_example()
-        .app_references(vec![54321, 11111, 55555, 22222])
-        .box_references(vec![
-            BoxReference {
-                app_id: 55555,
-                name: "b1".as_bytes().to_vec(),
-            },
-            BoxReference {
-                app_id: 54321,
-                name: "b2".as_bytes().to_vec(),
-            },
-        ])
-        .build()
-        .unwrap();
-
-    check_transaction_id(
-        &application_call_tx,
-        "GB4AYDJEHVBLOVSLCBOXG3KASTS3V6QV6GPB6F2BILG7L6J3P4OQ",
-    );
-    check_transaction_encoding(&application_call_tx, 169);
-}
-
-#[test]
-fn test_box_ref_missing_app_reference_encode() {
-    let application_call_tx: Transaction = TransactionMother::application_call_example()
-        .app_references(vec![54321])
-        .box_references(vec![
-            BoxReference {
-                app_id: 55555,
-                name: "b1".as_bytes().to_vec(),
-            },
-            BoxReference {
-                app_id: 54321,
-                name: "b2".as_bytes().to_vec(),
-            },
-        ])
-        .build()
-        .unwrap();
-
-    let result = application_call_tx.encode();
-
-    assert!(result.is_err());
-    let error_message = result.unwrap_err().to_string();
-    assert!(
-        error_message.contains("Box reference with app id 55555 not found in app references"),
-        "Expected missing app reference error, got: {}",
-        error_message
-    );
-}
-
-#[test]
-fn test_box_ref_missing_app_reference_decode() {
-    let encoded_tx_missing_app_ref = [
-        84, 88, 138, 164, 97, 112, 98, 120, 146, 130, 161, 105, 1, 161, 110, 196, 2, 98, 49, 130,
-        161, 105, 2, 161, 110, 196, 2, 98, 50, 164, 97, 112, 102, 97, 145, 205, 212, 49, 164, 97,
-        112, 105, 100, 205, 48, 57, 163, 102, 101, 101, 205, 3, 232, 162, 102, 118, 1, 163, 103,
-        101, 110, 167, 101, 120, 97, 109, 112, 108, 101, 162, 103, 104, 196, 32, 222, 189, 190,
-        157, 28, 11, 247, 214, 147, 68, 228, 226, 58, 211, 196, 121, 68, 26, 174, 253, 159, 1, 57,
-        38, 54, 88, 135, 169, 241, 177, 52, 144, 162, 108, 118, 205, 3, 231, 163, 115, 110, 100,
-        196, 32, 2, 204, 225, 113, 58, 8, 179, 189, 204, 74, 148, 128, 202, 244, 192, 188, 2, 202,
-        236, 227, 17, 198, 25, 62, 33, 204, 91, 40, 252, 44, 209, 74, 164, 116, 121, 112, 101, 164,
-        97, 112, 112, 108,
-    ];
-
-    let result = Transaction::decode(&encoded_tx_missing_app_ref);
-
-    assert!(result.is_err());
-    let error_message = result.unwrap_err().to_string();
-    assert!(
-        error_message.contains("Cannot find app reference index 1"),
-        "Expected missing app reference error, got: {}",
-        error_message
-    );
-}
-
-#[test]
-fn test_application_call_empty_value_encoding() {
-    let builder = &ApplicationCallTransactionBuilder::default()
-        .header(TransactionHeaderMother::example().build().unwrap())
-        .app_id(1234)
-        .on_complete(OnApplicationComplete::NoOp)
-        .to_owned();
-
-    let tx = builder.clone().build().unwrap();
-    let tx_with_empties = builder
-        .clone()
-        .approval_program(vec![])
-        .clear_state_program(vec![])
-        .args(vec![])
-        .account_references(vec![])
-        .asset_references(vec![])
-        .account_references(vec![])
-        .box_references(vec![])
-        .build()
-        .unwrap();
-
-    let expected_id = "AEAVEJUTYW5MFUWTDX6YPQS57FILUMVGDNYVB6ZGNNWL5Z4D43OA";
-
-    assert_ne!(tx, tx_with_empties);
-
-    // Because id's are a hash of the encoded bytes, we can be sure the encoded bytes are the same
-    check_transaction_id(&tx, expected_id);
-    check_transaction_id(&tx_with_empties, expected_id);
 }
